Refactor filesystem layout - reconsider how to group base versus wireless AP to improve future use cases
