+++ helpers/netbox/__init__.py
@@ -3,6 +3,5 @@ Base NetBox helper import script.  Import all objects from the classes,
 models, and functions that are present in the corresponding __init__.py
 "__all__" definition.
 """
-from .classes import *
-from .models import *
-from .functions import *
+from .base import *
+from .wireless_ap import *
