+++ helpers/netbox/wireless_ap/wireless_vars.py
