+++ requirements.txt
@@ -4,3 +4,5 @@ perflint ~= 0.7.3
 pydantic ~= 1.9.1
 pylint ~= 2.14.5
 pynetbox ~= 6.6.2
+requests ~= 2.28.1
+urllib3 ~= 1.26.11
